requirements.txt
grpcio==1.18.0
grpcio-tools==1.18.0
Pillow==5.4.1
protobuf==3.6.1
PyJWT==1.7.1
six==1.12.0
websockets==7.0
